part 7

+++ part4/blog-list/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (request, response) => {
 	const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -9,8 +10,12 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
 	const body = request.body;
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: "token missing or invalid" });
+	}
 
-	const user = await User.findOne({});
+	const user = await User.findById(decodedToken.id);
 
 	const blog = new Blog({
 		title: body.title,
@@ -28,6 +33,19 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: "token missing or invalid" });
+	}
+
+	const blog = await Blog.findById(request.params.id);
+
+	if (blog.user.toString() !== decodedToken.id) {
+		return response
+			.status(401)
+			.json({ error: "user not allowed to delete this blog" });
+	}
+
 	await Blog.findByIdAndRemove(request.params.id);
 	response.status(204).end();
 });
